+++ src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/EnemyBehaviour.cs
@@ -1,0 +1,7 @@
+using Code.Component;
+using Entitas.Generic;
+
+namespace Code
+{
+	public class EnemyBehaviour : ComponentBehaviour<GameScope, Enemy> { }
+}
