+++ src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/EnemyBehaviour.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f100bc976474484e828593d1de479c1f
+timeCreated: 1700384319
