+++ src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/SpawnActorBehaviour.cs
@@ -1,0 +1,7 @@
+using Code.Component;
+using Entitas.Generic;
+
+namespace Code
+{
+	public class SpawnActorBehaviour : ComponentBehaviour<RequestScope, SpawnActor> { }
+}
