--- src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/SpawnEnemyBehaviour.cs
@@ -1,7 +1,0 @@
-using Code.Component;
-using Entitas.Generic;
-
-namespace Code
-{
-	public class SpawnEnemyBehaviour : ComponentBehaviour<RequestScope, SpawnEnemy> { }
-}
