--- src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/SpawnEnemyBehaviour.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8d37bc1314f647659b31c1b5e891f1df
-timeCreated: 1700127827
