--- src/RerollKnight/Assets/Code/Boilerplate/Behaviours/Components/SpawnPlayerBehaviour.cs
@@ -1,7 +1,0 @@
-using Code.Component;
-using Entitas.Generic;
-
-namespace Code
-{
-	public class SpawnPlayerBehaviour : ComponentBehaviour<RequestScope, SpawnPlayer> { }
-}
