+++ src/RerollKnight/Assets/Code/Boilerplate/Features/BoilerplateFeature.cs
@@ -16,8 +16,7 @@ namespace Code
 			Add(new RemoveComponentsSystem<Clicked, GameScope>(contexts));
 
 			Add(new DestroyEntitySystem<CastAbility, RequestScope>(contexts));
-			Add(new DestroyEntitySystem<SpawnPlayer, RequestScope>(contexts));
-			Add(new DestroyEntitySystem<SpawnEnemy, RequestScope>(contexts));
+			Add(new DestroyEntitySystem<SpawnActor, RequestScope>(contexts));
 			Add(new DestroyEntitySystem<SetAllTargetsAvailability, RequestScope>(contexts));
 		}
 	}
