+++ src/RerollKnight/Assets/Code/Enemy/Components.cs
@@ -2,7 +2,5 @@ using Entitas.Generic;
 
 namespace Code.Component
 {
-	[RequestScope] public sealed class SpawnEnemy : FlagComponent, ICleanup<DestroyEntity> { }
-
 	[GameScope] public sealed class Enemy : FlagComponent { }
 }
