+++ src/RerollKnight/Assets/Code/Player/Components.cs
@@ -3,6 +3,4 @@ using Entitas.Generic;
 namespace Code.Component
 {
 	[GameScope] public sealed class Player : FlagComponent { }
-
-	[RequestScope] public sealed class SpawnPlayer : FlagComponent, ICleanup<DestroyEntity> { }
 }
