--- src/RerollKnight/Assets/Code/Player/SpawnPlayerSystem.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: c97067f1fcbf4806bc92bcdb1c298676
-timeCreated: 1686392638
