+++ src/RerollKnight/Assets/Scenes/SampleScene.unity
@@ -307,7 +307,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 256804607}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8d37bc1314f647659b31c1b5e891f1df, type: 3}
+  m_Script: {fileID: 11500000, guid: 1b9cdcecd9c54672993ed305daeefa2f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &256804611
